Add motor run (AFMS initialization broken)
added the lines to actually run the motors, a brake loop, and checking both sensors to make sure only one is activated before turning. I added the missing AFMS start line in setup, but it won't run since it isn't in the robot class. I tried putting it in, and it still didn't work.